app/requirements.txt
pytest==8.3.2
psycopg[binary]==3.2.9
python-dotenv==1.0.1